+++ fastinterp/fastinterp_tpl_partial_inline_arith_expr.cpp
@@ -150,7 +150,7 @@ struct FIPartialInlineArithmeticExprImpl
             }
             else
             {
-                DEFINE_CONSTANT_PLACEHOLDER_1(uint32_t);
+                DEFINE_CONSTANT_PLACEHOLDER_1(uint64_t);
                 rhs = *GetLocalVarAddress<OperandType>(stackframe, CONSTANT_PLACEHOLDER_1);
             }
         }
@@ -163,7 +163,7 @@ struct FIPartialInlineArithmeticExprImpl
             }
             else
             {
-                DEFINE_CONSTANT_PLACEHOLDER_1(uint32_t);
+                DEFINE_CONSTANT_PLACEHOLDER_1(uint64_t);
                 lhs = *GetLocalVarAddress<OperandType>(stackframe, CONSTANT_PLACEHOLDER_1);
             }
         }
@@ -177,7 +177,7 @@ struct FIPartialInlineArithmeticExprImpl
         }
         else
         {
-            DEFINE_CONSTANT_PLACEHOLDER_0(uint32_t);
+            DEFINE_CONSTANT_PLACEHOLDER_0(uint64_t);
             *GetLocalVarAddress<OperandType>(stackframe, CONSTANT_PLACEHOLDER_0) = result;
 
             DEFINE_BOILERPLATE_FNPTR_PLACEHOLDER_0(void(*)(uintptr_t, OpaqueParams...) noexcept);
