+++ fastinterp/fastinterp_tpl_partial_inline_assign.cpp
@@ -103,7 +103,7 @@ struct FIPartialInlineAssignImpl
         }
         else
         {
-            DEFINE_CONSTANT_PLACEHOLDER_0(uint32_t);
+            DEFINE_CONSTANT_PLACEHOLDER_0(uint64_t);
             rhs = *GetLocalVarAddress<OperandType>(stackframe, CONSTANT_PLACEHOLDER_0);
         }
