+++ fastinterp/fastinterp_tpl_partial_inline_rhs_assign.cpp
@@ -86,7 +86,7 @@ struct FIPartialInlineRhsAssignImpl
         }
         else
         {
-            DEFINE_CONSTANT_PLACEHOLDER_0(uint32_t);
+            DEFINE_CONSTANT_PLACEHOLDER_0(uint64_t);
             lhs = *GetLocalVarAddress<OperandType*>(stackframe, CONSTANT_PLACEHOLDER_0);
         }
